ShellTypes

Description

This module is used to define types that are going to be used by the shell. For example, CmdType is used to easily identify what needs to be evaluated by runIO

Documentation

data CmdType

CmdType allows us to know what type of command is being evaluated

Constructors

None

Pipe String [String]

Redir String